add dbcontext and test database connection

+++ ClinicManagementSoftware/src/ClinicManagementSoftware.Web/Endpoints/ProjectEndpoints/ToDoItemRecord.cs
@@ -1,5 +1,5 @@
 ﻿
 namespace ClinicManagementSoftware.Web.Endpoints.ProjectEndpoints
 {
-    public record ToDoItemRecord(int Id, string Title, string Description, bool IsDone);
+    public record ToDoItemRecord(long Id, string Title, string Description, bool IsDone);
 }
